+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.5.3
+- Improved error messages to provide suggestions on how to fix exceptions.
+- Added a Key check to block Bilocators with the same key from registering. This fixes Bilocators 
+trying to register BilocatorDelegates during a hot reload.
+- Changed Bilocators to register data from constructor instead of initState so the data is 
+available earlier.
+- Improved the demo app to test context.of
+
 ## 0.5.2
 - Improved readme.